refactor(layering): remove retired migration scaffolding
