+++ scripts/layering/record-runtime-mechanics-policy.ts
@@ -1,5 +1,5 @@
 import { parseSync } from 'oxc-parser';
-import { memberName, type ProductionSource, visitAst } from './cutover-policy-ast.ts';
+import { memberName, type ProductionSource, visitAst } from './layering-ast.ts';
 
 const DAEMON_RECORD_MECHANIC_CALLS = new Set([
   'runCmd',
